Rename Duke class/test class and related duke paths to Sherpass
As part of the team's product name
