--- src/main/java/seedu/duke/command/Command.java
@@ -1,10 +1,0 @@
-package seedu.duke.command;
-
-import seedu.duke.Storage;
-import seedu.duke.TaskList;
-import seedu.duke.Ui;
-
-public abstract class Command {
-
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
-}
