--- src/main/java/seedu/duke/constant/Indexes.java
@@ -1,22 +1,0 @@
-package seedu.duke.constant;
-
-public class Indexes {
-    public static final int SAVE_TASK_TYPE_INDEX = 0;
-    public static final int SAVE_TASK_MARK_STATUS = 1;
-    public static final int SAVE_TASK_DESCRIPTION_INDEX = 2;
-    public static final int SAVE_TASK_DATE_INDEX = 3;
-    public static final int DELETE_INDEX = 1;
-    public static final int OPTIONS_INDEX = 0;
-    public static final int MARK_INDEX = 1;
-    public static final int TASK_CONTENT_INDEX = 1;
-    public static final int HELP_OPTIONS_INDEX = 1;
-    public static final int TASK_DESCRIPTION_INDEX = 0;
-    public static final int TASK_DATE_INDEX = 1;
-    public static final int TASK_DESCRIPTION_INDEX_TODO = 1;
-    public static final int FIND_BY_TASK_DATE_INDEX = 1;
-    public static final int FIND_BY_TASK_DESCRIPTION_WITH_DATE_INDEX = 0;
-    public static final int FIND_BY_TASK_DESCRIPTION_NO_DATE_INDEX = 1;
-    public static final int FIND_BY_TASK_CONTENT_INDEX = 1;
-    public static final int DIRECTORY_INDEX = 0;
-
-}
