--- src/main/java/seedu/duke/exception/InputRepeatedException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exception;
-
-public class InputRepeatedException extends Exception {
-}
