--- src/main/java/seedu/duke/exception/InvalidInputException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exception;
-
-public class InvalidInputException extends Exception {
-}
