--- src/main/java/seedu/duke/task/Deadline.java
@@ -1,60 +1,0 @@
-package seedu.duke.task;
-
-public class Deadline extends Task {
-    protected String by;
-
-    /**
-     * Creates a constructor for the task 'deadline'.
-     * Deadline tasks accept both task description and date.
-     *
-     * @param description Task Description.
-     * @param by Task Date.
-     */
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by;
-    }
-
-    /**
-     * Returns task date.
-     *
-     * @return Task date.
-     */
-    @Override
-    public String getDate() {
-        return by;
-    }
-
-    /**
-     * Returns string version of task type.
-     *
-     * @return "D" for deadline.
-     */
-    @Override
-    public String getType() {
-        return "D";
-    }
-
-    /**
-     * Overrides existing task date with new date input.
-     *
-     * @param date Task Date.
-     */
-    @Override
-    public void resetInput(String date) {
-        by = date;
-    }
-
-
-    /**
-     * Returns a string version of the task content.
-     * Content includes mark status, task type， description and date.
-     * For ease of printing the task.
-     *
-     * @return Task content.
-     */
-    @Override
-    public String toString() {
-        return "[D] " + super.toString() + " (by: " + by + ")";
-    }
-}
