+++ src/main/java/seedu/sherpass/command/Command.java
@@ -1,0 +1,10 @@
+package seedu.sherpass.command;
+
+import seedu.sherpass.Storage;
+import seedu.sherpass.TaskList;
+import seedu.sherpass.Ui;
+
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+}
