+++ src/main/java/seedu/sherpass/exception/InputRepeatedException.java
@@ -1,0 +1,4 @@
+package seedu.sherpass.exception;
+
+public class InputRepeatedException extends Exception {
+}
