+++ src/main/java/seedu/sherpass/exception/InvalidInputException.java
@@ -1,0 +1,4 @@
+package seedu.sherpass.exception;
+
+public class InvalidInputException extends Exception {
+}
