+++ src/test/java/seedu/sherpass/SherpassTest.java
@@ -1,0 +1,12 @@
+package seedu.sherpass;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+class SherpassTest {
+    @Test
+    public void sampleTest() {
+        assertTrue(true);
+    }
+}
